Extract WalletActivator
